[Chore] #7 - Compositional Layout 파일 분리 및 세부 수정
